Removed turtle specific providers and services from DocumentContext.
