+++ src/document-context.ts
@@ -2,19 +2,12 @@ import * as n3 from 'n3';
 import * as rdfjs from "@rdfjs/types";
 import * as vscode from 'vscode';
 import { IToken } from 'millan';
-import { _OWL, _RDF, _RDFS, _SH, _SKOS, _SKOS_XL, rdf, sh } from '@faubulous/mentor-rdf';
+import { _OWL, _RDF, _RDFS, _SH, _SKOS, _SKOS_XL, sh } from '@faubulous/mentor-rdf';
 import { PredicateUsageStats, LanguageTagUsageStats } from '@faubulous/mentor-rdf';
 import { mentor } from '@/mentor';
 import { TreeLabelStyle } from '@/settings';
 import { DefinitionProvider } from '@/languages';
-import {
-	getIriLocalPart,
-	getIriFromIriReference,
-	getIriFromPrefixedName,
-	getIriFromToken,
-	getNamespaceDefinition,
-	getNamespaceIri
-} from '@/utilities';
+import { getIriLocalPart, getNamespaceIri } from '@/utilities';
 
 /**
  * A literal value with optional language tag.
@@ -70,21 +63,40 @@ export abstract class DocumentContext {
 	 */
 	readonly graphs: string[] = [];
 
-	private _tokens: IToken[] = [];
+	/**
+	 * Get the base IRI of the document that can be used for resolving local names into IRIs.
+	 */
+	baseIri: string | undefined;
 
-	private _baseIri: string | undefined;
+	/**
+	* Maps prefixes to namespace IRIs.
+	*/
+	namespaces: { [key: string]: string } = {};
 
-	private _namespaces: { [key: string]: string } = {};
+	/**
+	 * Maps resource IRIs to indexed tokens.
+	 */
+	namespaceDefinitions: { [key: string]: IToken } = {};
 
-	private _namespaceDefinitions: { [key: string]: IToken } = {};
+	/**
+	 * Maps resource IRIs to indexed tokens.
+	 */
+	references: { [key: string]: IToken[] } = {};
 
-	private _references: { [key: string]: IToken[] } = {};
+	/**
+	 * Maps resource IRIs to tokens of subjects that have an asserted rdf:type, including named individuals.
+	 */
+	typeAssertions: { [key: string]: IToken[] } = {};
 
-	private _typeAssertions: { [key: string]: IToken[] } = {};
+	/**
+	 * Maps resource IRIs to tokens of subjects that are class or property definitions.
+	 */
+	typeDefinitions: { [key: string]: IToken[] } = {};
 
-	private _typeDefinitions: { [key: string]: IToken[] } = {};
-
-	private _blankNodes: { [key: string]: IToken } = {};
+	/**
+	 * Maps blank node ids to indexed tokens.
+	 */
+	blankNodes: { [key: string]: IToken } = {};
 
 	/**
 	 * Information about the language tags used in the document.
@@ -173,80 +185,13 @@ export abstract class DocumentContext {
 	/**
 	 * Indicates whether the document is fully loaded.
 	 */
-	get isLoaded(): boolean {
-		return this._tokens.length > 0;
-	}
+	abstract get isLoaded(): boolean;
 
 	/**
 	 * Indicates whether the document is temporary and not persisted.
 	 */
 	get isTemporary(): boolean {
 		return this.uri.scheme == 'git';
-	}
-
-	/**
-	 * All tokens in the document.
-	 */
-	get tokens(): IToken[] {
-		return this._tokens;
-	}
-
-	/**
-	 * Get the base IRI of the document for resolving local names into IRIs.
-	 * @returns The base IRI of the document or `undefined`.
-	 */
-	get baseIri(): string | undefined {
-		return this._baseIri;
-	}
-
-	/**
-	 * Set the base IRI of the document for resolving local names into IRIs.
-	 * @param value The base IRI of the document.
-	 */
-	protected set baseIri(value: string | undefined) {
-		this._baseIri = value;
-	}
-
-	/**
-	* Maps prefixes to namespace IRIs.
-	*/
-	get namespaces(): { [key: string]: string } {
-		return this._namespaces;
-	}
-
-	/**
-	 * Maps resource IRIs to indexed tokens.
-	 */
-	get namespaceDefinitions(): { [key: string]: IToken } {
-		return this._namespaceDefinitions;
-	}
-
-	/**
-	 * Maps resource IRIs to indexed tokens.
-	 */
-	get references(): { [key: string]: IToken[] } {
-		return this._references;
-	}
-
-	/**
-	 * Maps resource IRIs to tokens of subjects that have an asserted rdf:type, including named individuals.
-	 */
-	get typeAssertions(): { [key: string]: IToken[] } {
-		return this._typeAssertions;
-	}
-
-	/**
-	 * Maps resource IRIs to tokens of subjects that are class or property definitions.
-	 */
-	get typeDefinitions(): { [key: string]: IToken[] } {
-		return this._typeDefinitions;
-	}
-
-	/**
-	 * Maps blank node ids to indexed tokens.
-	 */
-	get blankNodes(): { [key: string]: IToken } {
-		return this._blankNodes;
 	}
 
 	/**
@@ -280,32 +225,10 @@ export abstract class DocumentContext {
 	abstract getDefinitionProvider(): DefinitionProvider;
 
 	/**
-	 * Get the first token of a given type.
-	 * @param tokens A list of tokens.
-	 * @param type The type name of the token.
-	 * @returns The last token of the given type, if it exists, undefined otherwise.
-	 */
-	getFirstTokenOfType(type: string): IToken | undefined {
-		const n = this.tokens.findIndex(t => t.tokenType?.tokenName === type);
-
-		if (n > -1) {
-			return this.tokens[n];
-		}
-	}
-
-	/**
-	 * Get the last token of a given type.
-	 * @param tokens A list of tokens.
-	 * @param type The type name of the token.
-	 * @returns The last token of the given type, if it exists, undefined otherwise.
-	 */
-	getLastTokenOfType(type: string): IToken | undefined {
-		const result = this.tokens.filter(t => t.tokenType?.tokenName === type);
-
-		if (result.length > 0) {
-			return result[result.length - 1];
-		}
-	}
+	 * Event handler for when the document is changed.
+	 * @param e The document change event.
+	 **/
+	async onDidChangeDocument(e: vscode.TextDocumentChangeEvent): Promise<void> { };
 
 	/**
 	 * Get the prefix for a namespace IRI.
@@ -322,116 +245,16 @@ export abstract class DocumentContext {
 
 	/**
 	 * Maps blank node ids of the parsed documents to the ones in the triple store.
+	 * @deprecated This method is deprecated and will be removed in a future release.
 	 */
 	mapBlankNodes() { }
 
 	/**
-	 * Set the tokens of the document and update the namespaces, references, type assertions and type definitions.
-	 * @param tokens An array of tokens.
+	 * Sets the tokens of the document.
+	 * @deprecated This method is deprecated and will be removed in a future release.
+	 * @param tokens The tokens of the document
 	 */
-	setTokens(tokens: IToken[]): void {
-		this._tokens = tokens;
-		this._namespaces = {};
-		this._namespaceDefinitions = {};
-		this._references = {};
-		this._typeAssertions = {};
-		this._typeDefinitions = {};
-		this._blankNodes = {};
-
-		tokens.forEach((t: IToken, i: number) => {
-			switch (t.tokenType?.tokenName) {
-				case 'PREFIX':
-				case 'TTL_PREFIX': {
-					const ns = getNamespaceDefinition(this.tokens, t);
-
-					// Only set the namespace if it is preceeded by a prefix keyword.
-					if (ns) {
-						this.namespaces[ns.prefix] = ns.uri;
-						this.namespaceDefinitions[ns.uri] = t;
-					}
-					break;
-				}
-				case 'PNAME_LN': {
-					const uri = getIriFromPrefixedName(this.namespaces, t.image);
-
-					if (!uri) break;
-
-					this._handleTypeAssertion(tokens, t, uri, i);
-					this._handleTypeDefinition(tokens, t, uri, i);
-					this._handleUriReference(tokens, t, uri);
-					break;
-				}
-				case 'IRIREF': {
-					const uri = getIriFromIriReference(t.image);
-
-					this._handleTypeAssertion(tokens, t, uri, i);
-					this._handleTypeDefinition(tokens, t, uri, i);
-					this._handleUriReference(tokens, t, uri);
-					break;
-				}
-				case 'A': {
-					this._handleTypeAssertion(tokens, t, rdf.type.id, i);
-					this._handleTypeDefinition(tokens, t, rdf.type.id, i);
-					break;
-				}
-			}
-		});
-	}
-
-	private _handleUriReference(tokens: IToken[], token: IToken, uri: string) {
-		if (!this.references[uri]) {
-			this.references[uri] = [];
-		}
-
-		this.references[uri].push(token);
-	}
-
-	private _handleTypeAssertion(tokens: IToken[], token: IToken, uri: string, index: number) {
-		if (uri != rdf.type.id) return;
-
-		const subjectToken = tokens[index - 1];
-
-		if (!subjectToken) return;
-
-		const subjectUri = getIriFromToken(this.namespaces, subjectToken);
-
-		if (!subjectUri) return;
-
-		this.typeAssertions[subjectUri] = [subjectToken];
-	}
-
-	private _handleTypeDefinition(tokens: IToken[], token: IToken, uri: string, index: number) {
-		if (uri != rdf.type.id) return;
-
-		const subjectToken = tokens[index - 1];
-
-		if (!subjectToken) return;
-
-		const subjectUri = getIriFromToken(this.namespaces, subjectToken);
-
-		if (!subjectUri) return;
-
-		const objectToken = tokens[index + 1];
-
-		if (!objectToken) return;
-
-		const objectUri = getIriFromToken(this.namespaces, objectToken);
-
-		if (!objectUri) return;
-
-		const namespaceUri = getNamespaceIri(objectUri);
-
-		// TODO: Make this more explicit to reduce false positives.
-		switch (namespaceUri) {
-			case _RDF:
-			case _RDFS:
-			case _OWL:
-			case _SKOS:
-			case _SKOS_XL:
-			case _SH:
-				this.typeDefinitions[subjectUri] = [subjectToken];
-		}
-	}
+	abstract setTokens(tokens: IToken[]): void;
 
 	/**
 	 * Updates a namespace prefix definition in the document.
@@ -446,29 +269,6 @@ export abstract class DocumentContext {
 		delete this.namespaces[oldPrefix];
 
 		this.namespaces[newPrefix] = uri;
-	}
-
-	/**
-	 * Gets all tokens at a given position.
-	 * @param tokens A list of tokens.
-	 * @param position A position in the document.
-	 * @returns An non-empty array of tokens on success, an empty array otherwise.
-	 */
-	getTokensAtPosition(position: vscode.Position): IToken[] {
-		// The tokens are 0-based, but the position is 1-based.
-		const l = position.line + 1;
-		const n = position.character + 1;
-
-		return this.tokens.filter(t =>
-			t.startLine &&
-			t.startLine <= l &&
-			t.endLine &&
-			t.endLine >= l &&
-			t.startColumn &&
-			t.startColumn <= n &&
-			t.endColumn &&
-			t.endColumn >= (n - 1)
-		);
 	}
 
 	/**
