+++ src/languages/turtle/providers/turtle-prefix-completion-provider.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import { mentor } from '@/mentor';
 import { getPreviousToken } from '@/utilities';
 import { TurtleFeatureProvider } from '@/languages/turtle/turtle-feature-provider';
+import { TurtleDocument } from '../turtle-document';
 
 export class TurtlePrefixCompletionProvider extends TurtleFeatureProvider implements vscode.InlineCompletionItemProvider {
 	protected readonly prefixTokenTypes = new Set(["PREFIX", "TTL_PREFIX"]);
@@ -11,7 +12,7 @@ export class TurtlePrefixCompletionProvider extends TurtleFeatureProvider implem
 	}
 
 	provideInlineCompletionItems(document: vscode.TextDocument, position: vscode.Position, completion: vscode.InlineCompletionContext): vscode.ProviderResult<vscode.InlineCompletionItem[] | vscode.InlineCompletionList> {
-		const context = this.getDocumentContext(document);
+		const context = mentor.getDocumentContext(document, TurtleDocument);
 
 		if (!context) {
 			return null;
