+++ src/languages/turtle/services/index.ts
@@ -1,0 +1,1 @@
+export * from './turtle-prefix-definition-service';
