+++ src/languages/xml/xml-document.ts
@@ -4,6 +4,7 @@ import { mentor } from '@/mentor';
 import { DocumentContext, TokenTypes } from '@/document-context';
 import { DefinitionProvider } from '@/languages/definition-provider';
 import { XmlDefinitionProvider } from './providers/xml-definition-provider';
+import { IToken } from 'millan';
 
 /**
  * A document context for RDF/XML documents.
@@ -22,7 +23,7 @@ export class XmlDocument extends DocumentContext {
 	}
 
 	get isLoaded(): boolean {
-		return super.isLoaded && this.graphs.length > 0;
+		return this.graphs.length > 0;
 	}
 
 	public override getDefinitionProvider(): DefinitionProvider {
@@ -77,8 +78,12 @@ export class XmlDocument extends DocumentContext {
 		}
 	}
 
+	public override setTokens(tokens: IToken[]): void {
+		// TODO: Remove this method from DocumentContext as this is specific for the parser and document type.
+	}
+
 	protected getXmlBaseIri(data: string): string | undefined {
-		const regex =/xml:base="([^"]+)"/i;
+		const regex = /xml:base="([^"]+)"/i;
 		const match = regex.exec(data);
 
 		if (match) {
