+++ src/services/index.ts
@@ -1,4 +1,4 @@
 export * from './local-storage-service';
-export * from './prefix-definition-service';
+export * from '../languages/turtle/services/turtle-prefix-definition-service';
 export * from './prefix-downloader-service';
 export * from './prefix-lookup-service';
